Lower cleanup errors as warnings
Cleanups shouldn't make the whole upgrade fail - but instead we lower the messages as warnings.
Those could be printed out for e.g. if we run an upgrade meanwhile k3s is running - the upgrade still succeeds, but
unlikely we could restore the state as "ro" as meanwhile we might have workload running.