Find persistent on reset via udev data if its a devicemapper
During reset if we need to identify persistent and that its a
devicemapper, our usual methods wont work as ghw knows nothing about
devicemappers.

This provides a workaround limited in scope to the reset action and if
we have exhausted the other ways of getting the persistent partition and
tries to iterate over all dm- devices and get their info from udev
directly to match the label against the persistent label.

Signed-off-by: Itxaka <itxaka.garcia@spectrocloud.com>