:robot: Use earthly
Changes to use earthly instead of a Makefile.

As a side effect, it shows that some of the tests were not correctly
done and they were behaving differently due to the environment.

Plus as a side effect, becuase we are running as root in the earthly
env, we can test the full suite and drop the tests for non-root that
made no sense

Signed-off-by: Itxaka <itxaka.garcia@spectrocloud.com>