Fix shim name
Shim changed name when the grub2-efi packages had a different package
installed (shim-susesigned vs shim) and that resulted in a different
name for the shim file.

Package was updated by elemental-cli was not, so once we updated the
package on kairos, this resulted in a broken install.

Fix it by setting the old good name as shown in the package contents
(system/grub2-efi)

Signed-off-by: Itxaka <itxaka.garcia@spectrocloud.com>